content-cli/pyproject.toml
[project]
name = "daskalo-content-cli"
version = "0.1.0"
description = "Daskalo operator CLI — generates multimodal Greek lesson content using LangGraph + Vertex AI + Piper TTS"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "google-cloud-aiplatform>=1.142.0",
    "Pillow>=10.0.0",
    "google-cloud-firestore>=2.25.0",
    "google-cloud-storage>=3.10.1",
    "google-cloud-texttospeech>=2.22.0",
    "langchain-core>=1.2.21",
    "langchain-google-genai>=4.0.0",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
]

[project.scripts]
daskalo = "main:cli"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
